README.md
U N I is a new, independent lifestyle service that focuses on connecting people with similar passions, interests, and curiosities. We work with a vision of giving people of the community a way to pursue impulse fascinations and build upon existing talents and interests. The mentors of U N I are committed, enthusiastic, and certified by a U N I employee to guarantee satsifaction.